+++ src/documents/management/commands/document_sanity_checker.py
@@ -1,8 +1,5 @@
-import logging
 from django.core.management.base import BaseCommand
-from documents.sanity_checker import check_sanity, SanityError, SanityWarning
-
-logger = logging.getLogger("paperless.management.sanity_checker")
+from documents.sanity_checker import check_sanity
 
 
 class Command(BaseCommand):
@@ -15,13 +12,4 @@ class Command(BaseCommand):
 
         messages = check_sanity(progress=True)
 
-        if len(messages) == 0:
-            logger.info("No issues found.")
-        else:
-            for msg in messages:
-                if type(msg) == SanityError:
-                    logger.error(str(msg))
-                elif type(msg) == SanityWarning:
-                    logger.warning(str(msg))
-                else:
-                    logger.info((str(msg)))
+        messages.log_messages()